Homing Sequnece
Added Function 'home' to the driver code. The spool rotates slightly past a full rotation to find the home position. If it is not found an error message is displayed and the code will re-run to try to find the home position again. If found a success message is displayed.